Add margin to text area in TextEditor class
This commit adds a margin between the border of the window and the text area in the TextEditor class. The margin is created using an EmptyBorder that is added to the textArea. The size of the margin is specified by a new margin variable that can be adjusted to change the margin size. This change improves the visual appearance of the text editor by providing more space around the text area, making it easier to read and use.